Merge develop into master

See merge request baptiste72/sigl!62

+++ backend/authentication/views.py
@@ -1,7 +1,8 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.exceptions import AuthenticationFailed
-from rest_framework.permissions import AllowAny, IsAuthenticated
+from rest_framework.permissions import AllowAny
+from rest_framework.authentication import BasicAuthentication
 from rest_framework.authtoken.models import Token
 from rest_framework import status
 
@@ -38,10 +39,13 @@ class LoginView(APIView):
 
 
 class LogoutView(APIView):
-    permission_classes = [IsAuthenticated]
+    # permission_classes = [IsAuthenticated]
+    authentication_classes = [BasicAuthentication]
 
     def get(self, request):
-        request.user.auth_token.delete()
+        if request.user is not None:
+            if not request.user.is_anonymous:
+                request.user.auth_token.delete()
         return Response(status=status.HTTP_200_OK)
